requirements.txt
numpy
scipy
matplotlib
pyserial
pydobot